+++ protected/humhub/modules/dashboard/tests/config/test.php
@@ -1,10 +1,7 @@
 <?php
 
 return [
-    'fixtures' => [
-        'default',
-        'content' => 'humhub\modules\content\tests\codeception\fixtures\ContentFixture'
-    ],
+    'fixtures' => ['default']
 ];