+++ protected/humhub/modules/notification/tests/codeception/unit/rendering/MailTargetRenderTest.php
@@ -19,16 +19,4 @@ class MailTargetRenderTest extends HumHubDbTestCase
         $this->assertContains('<h1>TestedMailViewNotificationHTML</h1>', $renderer->render($notification));
         $this->assertContains('TestedMailViewNotificationText', $renderer->renderText($notification));
     }
-
-    public function testOverwriteViewFile()
-    {
-        $notification = notifications\TestedMailViewNotification::instance();
-        $notification->viewName = 'special';
-        $target = Yii::$app->notification->getTarget(MailNotificationTarget::class);
-        $renderer = $target->getRenderer();
-        $result = $renderer->render($notification);
-        $this->assertContains('<div>Special:<h1>TestedMailViewNotificationHTML</h1></div>', $result);
-        $this->assertContains('<div>MyLayout:', $result);
-        $this->assertContains('TestedMailViewNotificationText', $renderer->renderText($notification));
-    }
 }