+++ protected/humhub/modules/stream/widgets/views/stream.php
@@ -100,6 +100,6 @@ $this->registerJsVar('defaultStreamSort', ($defaultStreamSort != '') ? $defaultS
 
 <!-- show "Load More" button on mobile devices -->
 <div class="col-md-12 text-center visible-xs visible-sm">
-    <button id="btn-load-more" class="btn btn-primary btn-lg "><?php echo Yii::t('ContentModule.widgets_views_stream', 'Load more'); ?></button>
+    <button id="btn-load-more" data-action-click="loadMore" data-action-target="#wallStream" data-ui-loader class="btn btn-primary btn-lg "><?= Yii::t('ContentModule.widgets_views_stream', 'Load more'); ?></button>
     <br/><br/>
 </div>