+++ protected/humhub/modules/user/tests/config/test.php
@@ -1,10 +1,11 @@
 <?php
 
 return [
-    'fixtures' => ['default',
+    'fixtures' => [
+        'default',
         'user_follow' => 'humhub\modules\user\tests\codeception\fixtures\UserFollowFixture',
-        'user_mentioning' => 'humhub\modules\user\tests\codeception\fixtures\UserMentioningFixture',
-        'content' => 'humhub\modules\content\tests\codeception\fixtures\ContentFixture']
+        'user_mentioning' => 'humhub\modules\user\tests\codeception\fixtures\UserMentioningFixture'
+    ]
 ];